opt(compiler): eliminate redundant nested casts (as same-type, Math.imul, mixed binary)

+++ test/cases/phase0/math/imul/expected.c
@@ -3,6 +3,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%d\n", (int32_t)((int32_t)(3) * (int32_t)(4)));
+    printf("%d\n", 3 * 4);
     return 0;
 }
